requirements.txt
uvicorn
numpy
fastapi
pydantic
tiktoken
openai
asyncio
datetime
sse_starlette
pydantic_settings
uuid
tqdm
pdfminer
redis
poetry
requests
pytest
gunicorn
sendgrid
azure.servicebus
azure.storage.blob
azure.keyvault.secrets
azure.identity
logging
python-dotenv